[HeaderBar] refactoring and optimalize: spliting component and make code more readeable
