+++ components/LandingPage/HeaderInfoBar/HeaderBarComponent/IconsSection/IconsSection.component.tsx
@@ -4,7 +4,7 @@ import s from "./IconsSection.component.module.scss";
 function IconsSection({ icon, link }: { icon: StaticImageData; link: string }) {
   return (
     <a href={link} target="_blank">
-      <Image className={s.icon} src={icon} alt="icon" width={30} height={30} />
+      <Image priority className={s.icon} src={icon} alt="icon" width={30} height={30} />
     </a>
   );
 }
